Finished implementation.
